Bugfix: imports within a package could violate rules
Imports-with-a-package were allowed so that one could specify

	a.b.c=a.b

(a package is not allowed to import from its sibling packages) and
it would still be allowed to import internally (that is, within
a.b.c, you should be able to say "from a.b.c.d import e").

But it was done too broadly, and created a situation where one
cannot isolate a piece within a package; that is, a rule

	a.b=a.b.c

(no importing from a.b.c in a.b) was not enforced at all.